+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Unreleased
 
+- add `NumBytes` to `PrimeField`
+- rename package to `@nocturne-xyz/crypto-utils`
+- add module `hybrid-encryption` containing a generic `HybridCipher` 
+- add separate `fromEntropy` method to `PrimeField` for reducing arbitrary `Uint8Array`s into field elements
+- add `toBytes` and `fromBytes` to curve
+- make `ZModPField.toBytes` and `ZModPField.fromBytes` algorithmic constant-time
 - add separate best-effort constant-time scalar mul
 - restructure into two sub-modules: `hashes` and `algebra`.
 - impl Baby JubJub curve using `TwistedEdwardsCurve` abstraction